Fix autocomplete fields initialization
While using django-admin-bootstrapped and django-autocomplete-light
there are some cases where an empty hidden inline form (that for adding
a new inline) has an autocomplete field with id that contains a counter
instead __prefix__ string, which is incorrect for
django-autocomplete-light initializer. Therefore, I took advantage of
the fact that name of such a field (in an empty hidden inline form)
always includes the __prefix__ string and changed the implementation of
the django-autocomplete-light initializer.